Add block_ngram to fast translate, rename test_beam_search, fix semantically wrong tests.
